[refactor]리팩토링
