+++ src/main/java/step3/domain/RandomMoveStrategy.java
@@ -1,4 +1,6 @@
-package step3;
+package step3.domain;
+
+import step3.util.RandomUtil;
 
 public class RandomMoveStrategy implements MoveStrategy {
